Build a function signature object

Bundles argument types and an optional return type into a single object that can be applied to one or more functions via with_signature(). Useful when several functions share the same shape.

Usage

signature(..., .return = NULL)

Arguments

...

Named type specifications (one per argument).

.return

Return-type specification, or NULL.

Value

A type_signature object.

See also

Examples

sig <- signature(x = "numeric", y = "numeric", .return = "numeric")
add <- with_signature(function(x, y) x + y, sig)
add(2, 3)
#> [1] 5